Grow-IQ/message/message/message/requirements.txt
# Core FastAPI dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database and ORM
sqlalchemy==2.0.34
alembic==1.13.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Data validation
pydantic==2.10.4
pydantic[email]==2.10.4

# WebSocket support
websockets==12.0

# Environment and configuration
python-dotenv==1.0.1

# Development and testing
pytest>=8.2.0
pytest-asyncio>=0.24.0
httpx>=0.27.0

# Production dependencies (optional)
gunicorn==21.2.0
psycopg2-binary==2.9.9  # For PostgreSQL
